(GH-1167) Honor run_plan(..., _run_as => ...) from apply()
Previously, apply() blocks would ignore the `_run_as` argument passed to
their containing plan.

    plan a() {
      run_plan(b, _run_as => 'someone')
    }
    plan b() {
      apply(...) { }
    }

In this case, the catalog compiled for the `apply()` block in plan b
would be executed as whatever user bolt connected to the user as instead
of the "someone" user specified via `_run_as`. This happened because the
Applicator class failed to replicate the full executor behavior of
running a task.

The Applicator now honors `run_as` if it's set on the executor, passing
it through to the transport's `batch_task` method.

This already worked fine for `run_as` set on a Target in the inventory
or for the `--run-as` flag.