(GH-1205) Allow powershell task parameters with `type` in the name
Previously, when parsing xml in the `Get-ContentAsJson` implementation a regex used to filter out xml elements with whos `MemberType` `Properties` had a name that was preceeded with `__` or contained the string `type`. Filtering on the string `type` caused task parameters that contain that string to be ignored, preventing users from being able to effectively pass powershell task paramters with the `type` in the name. Unfortunately, with the current implementation powershell parameters with the name `type` will be clobbered by a `Property` called type. However this commit updates the regex to *only* filter out the property that exactly matches `type`. Note that this work allows powershell task parameters with `type` in the name, but not the name `type` and only affects powershell < 2.

closes https://github.com/puppetlabs/bolt/issues/1205