Include 'Creating a project' in the title
The installation doc is not just installation, but also has an overview of what a project directory looks like.  I already had bolt installed, so I skipped this and ran into problems after I set up the nodes.  This makes the title a little more cumbersome, but describes the other two thirds of the document for people who skim.