(ORCH-2341) Allow for multi-target actions in bolt_server
Refactors the /ssh/* and /winrm/* endpoints in bolt_server to support
future additions of actions that accept multiple targets at once.

To pass the JSON schema validations for SSH or WinRM transport details,
a bolt_server request may now have either a single `target` property, or
an array of `targets` (but not both). Other existing schemas for
individual action validations are already responsible for validating the
use of the `target` key; they will not begin to accept `targets`.
`targets` is intended for future use in ORCH-2341.