(GH-1158) Quote paths for interpreter and task when using winrm
Previously when running a task over winrm where a space was contained in a user defined `interpreter` path or the path to the temporary task executable the task could fail due to not quoting the paths to handle the spaces. This commit updates the winrm transport to quote both paths when `interpreters` is configured. Note that for the `interpreter` path, if the user has specified a quoted path, no additional quoting will be added.

closes https://github.com/puppetlabs/bolt/issues/1158