(GH-1174) Expand file paths for orchestrtor client config
Previously when the `token-file` or `cacert` file paths were used to construct a new orchestrator client instance, the paths were not expanded. Other paths in bolt expand and there are examples in the wild where paths need to be expanded (for example `token-file: ~/.puppetlabs/token`). This commit updates the orch transport to expand file paths before creeting an Orchestrator client object.