(ORCH-2317) Update upload_file in bolt-server to work with single files
Prior to this commit, the bolt-server implementation of upload_file would
always attempt to download with a source location of cache_dir/job_id (i.e. a
top-level directory). This breaks downloads for single files, since a single
file download request could have a destination with a fully qualified
filename, in which case the download would make the destination filename in to
a directory and download the single file in to that directory.

This commit updates upload_file in transport_app.rb to special case the single
file scenario, so that the upload source is `cache_dir/job_id/relative_path`
(i.e. a fully qualified path to the single file).