(maint) Standardize params vs. options and use symbols
Previously, we split out params and options into separate hashes
based on whether the keys started with _, but then left the keys that
way. That meant that the internals of Bolt were littered with code that
included leading underscores and thus had some evidence of where those
options originated from. As the leading underscore is an implementation
detail of the Puppet functions, we now consistently strip them out when
exiting the Puppet language and standardize them to symbols.