(maint) Eagerly evaluate OS in analytics client
Previously, we spun evaluation of the client OS into another thread to
avoid slowing down the main execution thread by needing to require
facter. This caused a race condition in facter between the thread
compting the OS and the main thread loading PAL which also needs to
evaluate some facts. As it turns out, by the time the analytics client
is instantiated we've already required facter anyway, so we don't have
to pay that cost. Evaluating the OS fact is negligible (~25ms), so we
now do it in the main thread to avoid the race condition.