(GH-1152) Raise helpful error message when tash metadata has wrong form
Previously it was assumened the the `requirements` field in a task's `implementations` metadata would be an array. When a `String` is specified (an easy mistake to make) a cryptic error message was surfaced. This commit adds some error checking and alerts the user that the metadata has an error in it.