(GH-1245) Standardize JSON and Puppet Result objects
This commit standardizes the Result object interface in plans with the JSON representation. This involved exposing a new `status` method on the Puppet Type, removing the `node` key from the status hash (in favor of `target`), and renaming `result` to `value`. This change is able to be toggled on and off with the newly introduced `$future` global variable.