(GH-1147) Allow resolve_reference plugins everywhere in inventory
Previously, resolve_reference plugins could only be used in the
inventory in select ways. Namely, they could only be used to lookup
targets or to set config on a group or target. They couldn't be used to
set any other attributes of a group (vars, facts, etc) or to set the
arguments to another plugin.

We now allow every value in the inventory to be specified using a plugin
reference, including the inventory itself. Keys are never allowed to be
plugin references.

Some references will be resolved as soon as the inventory is loaded.
For a group, these are `name`, `targets`, `groups`. For each target, we
will immediately resolve its `name`, uri`, and `alias`.

The other fields (`vars`, `facts`, `features`, `config`, `plugin_hooks`)
are termed "group/target data" and are resolved only when looking up a
target with that data.

We no longer use `Concurrent::Delay` objects to defer plugin evaluation.
Instead, the group tracks whether its data has been resolved and will do
it on demand if necessary. Similarly, the group tracks which of its
targets have been resolved and will resolve them as needed.

References are now allowed to return data containing more references,
which will be recursively re-evaluated until all the references are
gone. The exceptions to this rule are the `targets` and `groups` keys,
which will only be re-evaluated until the top-level object is no longer
a reference. This allows them to return references for their "data"
which will still be lazily resolved.

We now use an internal `Bolt::Inventory::Target` object to track the
current state of a target in the inventory and reconcile it with the
data set by its groups and config. Previously, this was a combination of
a hash on the inventory as well as the `Bolt::Target2` class. The
`Bolt::Inventory::Target` is now the canonical representation of the
current state of a target, tracking both the data set on that target
individually as well as querying the data from its group.

Instead of computing and caching the "final" merged data for the target,
we now only store the target-specific data and the cached group data
separately and compute results when needed. This allows us to invalidate
the group cache less frequently, as it now needs only be done when the
target is added to a new group and not whenever any value is set on the
target.

The `Bolt::Target2` class is now a thin, stateless wrapper around the
inventory and defers to the internal `Bolt::Inventory::Target` object
for all of its logic. This decouples the way we expose targets to plans
from the way that we store their state internally.

This update to the way we store target data simplifies the evaluation
patterns for the references in a group. Specifically, we query the
target-level data in each group exactly once, when we initially reify
the target (add it to the inventory via `get_targets`). After that, the
data is no longer needed as it is stored in its own space on the target.
We query the group-level data only when the target is first reified or
when its list of groups changes.