(GH-1211) Return Target object with plan function `add_facts`
In order to be more consistent with the other `set_*` plan methods for setting target data, this commit updates the `add_facts` function to return a `Target` where previously a `Hash` of the updated facts was returned. Given this is a breaking change it is gated with the `future` flag.