(docs) Fix absolute links and correct puppetdb config location
This fixes some links which where hardcoded to in-repo markdown docs to
be relative so they link to the production site pages. This also
corrects a misprint that the only config location for puppetdb
connection is `~/.puppetlabs/bolt`, when it is actually the Bolt config
file you're currently using (regardless of location).