(GH-1322) Validate plugin config against task metadata
Previously when a config section existed in a `bolt_plugin.json` specification any config that was set in `bolt.yaml` would be passed to the task as under the key `_config`. Tasks that expect values to be set in either `bolt.yaml` or `inventory.yaml` would have to handle munging the values gathered from either file. This commit begins to phase out the config section in `bolt_plugin.json`. If there is not a config section in that file config values are validated against the intersection of the parameter metadata for the tasks that implement each hook. Values set in `bolt.yaml` are merged with those in inventory.yaml and passed as parameters. No `_config` is passed along.