(maint) Install modules from bolt's Puppetfile for acceptance tests
Previously for component acceptance tests once bolt is cloned from git and a gem is built, a puppetfile was created in the default moduledir containing a reference to module versions independed from the modules shipped with bolt. We had to remember to manually update both the bolt Puppetfile as well as the acceptance test setup script when updating module references. With this commit, we use the Puppetfile from the bolt repo as the sole source of modules to be used in component acceptance tests.