(GH-1345) Install bundled modules on gem path in acceptance test
Previously we maintained a subset of modules needed for acceptance tests by statically defining module versions in acceptance test setup. This is prone to error because when modules are updated in the Puppetfile, there is a risk the acceptance test will not be updated. Also, the statically defined modules were installed into the default boltdir modulepath, not the modules dir in the gem path. Now that we ship the Puppetfile with the bolt gem, we can use it to simply install the modules in the expected bolt gem path with r10k to mirror modules shipped in system packages.