(maint) Raise error when hiera data is empty, plus docs fixes
This includes a few small improvements:
1. When using the [Boltdir
repository](https://github.com/puppetlabs/Boltdir) we provide an empty
hiera config. If left in place Bolt will try to load this which results
in a traceback when we try to access data from the file. This PR adds a
check which will ignore the hiera config if there's no data in it.
2. When using `--configfile` as a command line argument the parent
directory of the specified file is used as the Boltdir to load other
Bolt config, such as the inventory. This PR adds documentation to the
flag letting users know we do this.
3. Lastly, we had some plain URLs in the `using_plugins.md` file, which
in github markdown will render as links but in Dita don't get rendered
as links. This makes them have explicit links syntax, so they render as
links on the website.