(GH-1162) Expand configured paths relative to Boltdir
There are a number of config options for Bolt that accept filepaths.
Previously we expanded the path relative to where Bolt was run in some
cases, and relative to the Boltdir in others. This standardizes all
configurable paths to be expanded relative to the Boltdir if they are
defined in a config file, and relative to the current working directory
if they are specified on the CLI. This change doesn't affect absolute
paths.  This provides consistent and predictable behavior for users. New
path expansions are gated on the `future` option being set to `true`.

Closes #1162