(maint) Add error handling for invalid BOLT_INVENTORY
Previously when a user would make the mistake (easy to do) of specifying a file path for `BOLT_INVENTORY` a stack trace would be surfaced describing an argument error for the `Bolt::Inventory::ValidationError` initialize method. With this commit we ensure that the parsed BOLT_INVENTORY from environment is a hash and that if an unsupported version is detected the ValidationError is initialized with the expected arguments.