(GH-1361) Add Singular version of TargetSpec
TargetSpec accepts a single Variant[String, Target] OR an Array of the
same, and is understood to best be treated as if it could refer to
multiple endpoints.

Sometimes a plan (or task) parameter should accept a single target only.
At present, there is no built-in data type to support that.

This commit creates and provide a data type defining a single
Target(Spec): Boltlib::SingleTargetSpec.

Aliased to `SingleTargetSpec` to let users avoid typing "Boltlib", just
like the others.

Fixes #1361