(GH-1200) Print json instead of yaml with inventory show --detail
Previously nested hashes appear to have been converted to yaml anchors when using the --detail flag with inventory show. In order to avoid issues with printing yaml, this commit updates the human outputter to print JSON instead of yaml. Additionally there was a bug in the intial implementation where, in the Target{2} detail method the `alias` data was attempted to be collected from an instance variable `@target_alias` which did not exist. This resulted in a nil value for that every time that needed special escaping in the yaml print. This commit updates Inventory`target_alias` method that collects an array of aliases that point to a target's name in inventory. Additionally, `target_alias` was not fully working (especially when a new target was created in a plan), this commit updates Inventory2 to handle target aliases. With this addition it should be possible to copy/paste target data from the output of target show to an inventoryfile to have a valid target.