(GH-1350) Configure plugins using other plugins
Previously, the `plugins` section of `bolt.yaml` was required to be
completely static. Now, plugin can use the same `_plugin` hashes as in
`inventory.yaml` to define their configuration. For example, this allows
secrets for plugins to be encrypted or read from vault.

We now instantiate (and thus validate) all plugins that are configured
in `bolt.yaml` at the time the plugin system is initialized.