Don't use update_from_file to update_from_inventory
Previously, the update_from_inventory method was implemented in terms of
update_from_file and relied on the fact that most of the behavior of
update_from_file would be a noop as it operated on data disallowed in
the inventory. This commit extracts the common behavior, which is only
related to configuring transports, and calls that method from both
places. This allows the behavior for the config file to be separate from
behavior for config from inventory.