(GH-1350) Allow plugin_hooks to contain references
Previously, the value of `plugin_hooks` had to be static with all
hooks and parameters being defined in the config file. We now allow
these hooks and parameters to be looked up using plugins, which is
helpful if an external source is responsible for controlling the Puppet
version to use or if the Puppet install method requires the knowledge of
some secret.