(GH-1377) Add project migrate command
This adds support for a `project migrate` command that migrates a Bolt
project to the newest version. Specifically, the inventory file for the
project will be migrated from an earlier version (v1) to the latest
version (v2). Changes to the inventory file are made in place and
comments/formatting are not preserved during migration.