(GH-1269) Add password-prompt CLI flags and deprecation warning for optional param
This commit adds two new CLI flags: `password-prompt` and `sudo-password-prompt`. In order to make parsing commands simpler, moving forward all CLI options either expect a mandatory parameter or do not accept a parameter at all. The `password` and `sudo-password` options violate this by allowing users to optionally set the value on the CLI which, when no value is set a prompt is displayed and the user can enter a password. Currently we will only be warning when the user does not supply a value for `--password` or `--sudo-password`. In 2.0 supplying a value will cause an error.