(GH-1350) Allow entire `plugin_hooks` to be set with a reference
Previously, setting `plugin_hooks` to a `_plugin` reference would cause
a failure because we would merge the default `plugin_hooks` with the
reference before evaluating it. That caused the reference to receive
arguments it didn't expect. We now evaluate the configured
`plugin_hooks` key and then merge it over top of the defaults afterward.