Break travis tests into a matrix
Previously, we ran a set of steps in a single job in travis:
* Rubocop
* Most rspec tests
* Rspec tests for embedded modules
* Docs validation
* Expensive rspec tests that use external services

This took ~17-18 min on average and resulted in a huge amount of output
to sort through to see what failed.

We now run each of those steps in a separate cell of a matrix job. That
makes it easier to quickly see what failed and it also cuts a lot of
time off the test runs.

This adds a `travis` namespace to the Rakefile, with separate
`travis:<check>` jobs for each cell in the matrix. That gives us an easy
ways to define new steps and to manage what those steps should do, while
separating the set of jobs that travis care about from the set of jobs
that developers care about while working locally.

We also now run `rubocop` and `docker-compose build` with `--parallel`,
speeding each of those steps up a bit.