Always return boolean from file predicate functions
Previously, file::exists and file::readable could return nil in the case
where the file path was specified relative to a module and didn't exist.
That would cause the function (and in turn the plan) to fail with an
invalid return type.

We now ensure these functions always return a boolean.