Stub analytics client during tests
Setting up the analytics client (even the noop client) normally requires
computing the set of bundled content, which is a costly operation. We
now explicitly stub the client out to ensure we don't need to compute
that.

This also removes one analytics integration test that is duplicated by
existing unit tests for the `bundled_content` and
`report_bundled_content` methods.