Don't parse params if there aren't any params
Bolt can accept params in two ways: `--params <json blob>` or
`param1=value1 param2=value2`. If the latter method is used, we load the
task/plan metadata and try to JSON parse the parameter values, using the
specified parameter types to determine whether the values are valid.

Previously, we determined whether to do that based on whether `--params`
was the method being used. If the user didn't pass `--params`, then we
would load the task/plan and try to parse and validate params. However,
in the case where _no_ params were specified at all, we still loaded the
task/plan and then did basically nothing with it.

Because loading a task/plan is expensive and we happen to run plans
without parameters rather frequently in tests, this commit updates this
logic to skip parsing params if no params were supplied. That cuts 10-20
seconds off the runtime of the test suite.