(GH-1382) Ignore settings for localhost target when running plugins
Previously, we passed the Bolt config through to the "empty" inventory
used by plugins. That meant that any bolt.yaml configuration for the
`local` transport and any CLI flags were used when running plugins.
For example, running bolt with `--run-as` would cause plugins to run as
the `run-as` user, which is clearly not intended.

We now bypass the standard Bolt config and just use the defaults,
which means that plugins will always run against a standard unconfigured
localhost transport.