(GH-1409) Show plan info using puppet strings
Previously, the info we showed with `bolt plan show <plan>` was very
basic and only included some of the information that Puppet could load
from the plan object. We now use puppet strings to parse the docstrings
for the plan so we can show description for the plan and its parameters
as well as parameter defaults.

We also now show the same data for YAML plans, though not through Puppet
strings.