(GH-1244) Serialize plan variables as pcore
There are 4 custom types in the Bolt plan language that we want to make
available inside apply blocks: Target, Result, ResultSet, ApplyResult.
For each of the \*Result\* objects this is accomplished by serializing the
plan variables to the apply block as pcore instead of as json. We also
needed to add pcore_init methods to the implementation classes of the
objects, because the pcore deserializer will by default call
`initialize` on the class and pass in the values of the attributes of
the type, which doesn't match the signature of the actual
implementations. Using `_pcore_init_hash` and `_pcore_init_from_hash`
lets us define what data gets serialized and how it gets deserialized
for those types.

Target objects are trickier for a few reasons:
1. They must have an associate inventory
2. They should not be modified or created inside an apply block

Instead of trying to create the right guard rails around using Target
objects in apply blocks and pass around an inventory, it's easier to
think of (and implement) targets and inventory in apply separately, as
ApplyTarget and ApplyInventory. When creating the catalog compiler we
create an ApplyInventory, which implements errors for invalid apply
functions and has ApplyTarget as it's implementation class. This
protects the inventory within apply blocks, and clearly separates the
concept of Targets in plans, and Targets in apply blocks.

Closes #1244