(GH-1427) Use environment variable for puppetdb config path in Windows
This updates how Bolt locates the puppetdb config file on Windows.
Previously, the path was hardcoded as `C:\ProgramData\`, which could
cause issues if the user redirected the folder or had a different system
drive. Bolt will now use the environment variable
`%CSIDL_COMMON_APPDATA%` to get the known folder.