(maint) Search plan subdirectories for plan filename
We now have to discover the language a plan is written in based on it's
filename, which requires constructing the filepath from the plan name
(i.e. `foo::bar::baz`) and testing if the file exists with either `.pp`
or `.yaml` extensions. This search falsely assumed that the plan path
could have at most 2 values: the modulename and the plan filename.
Instead it can have the modulename, and a path relative to the `plans`
directory. This modifies the logic to correctly load nested plans.