(GH-1208) Ensure Puppet agent version with apply_prep
This updates the `apply_prep` plan function to ensure the specified
version of the Puppet agent is installed on the target. Previously, the
`apply_prep` function would only install the agent if the target did not
have any version of the agent installed and did not have the
`puppet-agent` feature set. Now, `apply_prep` will always install the
specified version of the agent on the target, including when an
already-installed agent would be upgraded or downgraded to another
version.