(maint) Fix bolt_server file_cache spec test
The file cache test sets a purge timeout/interval to 1 second, then sleeps for 2 seconds to allow enough time for the purge timer to fire. In most cases this is enough time for the mocked mutex to recieve `with_write_lock` one time. In some cases the timer fires twice. Given the interval is 1 second and we wait for 2 this is not unreasonable. This commit expects the timer to fire at most 2 times in the 2 seconds.