(GH-1500) Support .yml for bolt config and inventory
This adds support for the `.yml` extension for bolt config and inventory
files. Bolt will recognize either `bolt.yaml` or `bolt.yml` as the root
of a local Boltdir, with `bolt.yml` following the same behavior as
`bolt.yaml` with respect to project directory type precedence. If both
`.yaml` and `.yml` files are present for config or inventory, `.yaml` is
preferred. Any yaml files generated by bolt still have the `.yaml`
extension.