(ORCH-2290) Handle Java Uncloneable exception in Bolt::Util#deep_clone
When attempting to `deep_clone` an instance of a `Bolt::Config` class with a custom Logging appender that references a Jruby/Clojure logger as an instance variable an exception (`Java::JavaLang::CloneNotSupportedException`) is raised that is not handled by rescuing `TypeError`. This commit updates the method to handle `CloneNotSupportedException`s the same way `TypeError` exceptions are handled.