(GH-1431) Add support for trusted external facts
This adds support for the new trusted external facts command. It adds a
configuration option `trusted-external-command` where users can
configure a path to their trusted external facts executable. We pass that
value through to the Puppet settings config of the same name. This
allows users to access trusted external facts in `$trusted['external']`.
This feature is experimental in both Puppet and Bolt, and this API may
change or be removed.

Closes #1431