(GH-1159) Rename bolt-server docker artifacts
Having the bolt-server Dockerfile at the root of the Bolt repo is
confusing for users who are doing the hands-on lab and find this
Dockerfile by mistake instead of the hands-on lab one. This renames
Dockerfile to Dockerfile.bolt-server and docker-compose to
docker-compose-bolt-server in order to clarify that they're used for
bolt-server and not bolt.