(GH-1319) Execute in run-as user's home dir when using run-as with SSH
Previously, connecting as one user and running as a different user via
run-as would cause the operation to execute in the home directory of the
connected user instead of the run-as user. This difference doesn't
usually matter because the common run-as case involves escalating to
root, in which case root can read the directory anyway. However, when
connecting as root and then *dropping* to a less-privileged user, this
would leave the process in a situation where it might not have read
access to its working directory. This problem also arises when using NFS
with the standard root squash configuration, whereby root still can't
read the home directories of non-root users.

We now prepend a `cd` command with no argument before every command we
execute when using run-as. This form of `cd` implicitly changes to the
value of $HOME. We also now run with `sudo -H` which instructs sudo to
reset $HOME to the home directory of the user we're switching to.

This issue caused particular problems for the libexec tasks used to
apply Puppet code. Those tasks use Puppet::ModuleTool::Tar which in some
cases will use `Dir.chdir` with a block while extracting files. This
form of `Dir.chdir` will change to a temporary directory to do its work,
and then will attempt to change back to the directory it was in
previously. In cases where that directory is unreadable by the current
user, it would fail to switch back and the task would error. We now
ensure that those tasks will be running from a current working directory
where they are allowed read access.