(GH-1338) Import node definitions when compiling an apply block
During ordinary compilation, definitions are stored on the parser at
parse time and then added to the Program node at the root of the AST
before evaluation. Because the AST for an apply block has already been
parsed and is not a complete tree with a Program at the root level, we
need to rediscover the definitions and construct our own Program object.
This step is only necessary for apply blocks in plans, not when using
`bolt apply` as the AST is already a Program in that case.