(PE-27909) Return plan metadata for `init` plan
Previously when the plan metadata is requested for the special `init.pp` plan (with the same name as its parent module) the metadata endpoint would not trim the `init` from the name before calling in to get_plan_info. With this commit the endpoint has been updated to handle plans referenced only by `[module name]::init`.