(GH-1493) Replace Task struct with class
The previous use of Struct here didn't have many benefits, as it's
basically a proper Class. It was also sloppy about what it accepted as
arguments, as the metaprogramming involved made it possible to use both
keyword or string keys, and made every parameter effectively optional.

This is now written as an ordinary class with a positional initialize
method. For the common case where we have a TaskSignature object that we
want to turn into a Bolt::Task, there is now a from_task_signature
factory method. This simplifies the code and makes it easier to keep
track of data.