(GH-1493) Use Bolt::Task when showing task info
Previously, the get_task_info method used by `bolt task show` expected
to receive a task hash generated from a TaskSignature object. As we now
have a Bolt::Task class, it makes sense to use that as the standard
interface. This also ensures that whether we're inspecting or running a
task, we always have a Bolt::Task object to work with.