(GH-1544) Remove empty strings and objects from results in human output
This changes the human outputter so that it does not display empty
strings or empty JSON objects. Previously, if a result only had an
`_output` key, the human outputter would dislay the value of `_output`,
even if it was an empty string or whitespace. If the only key present
was `_output`, the outputter would also display an empty JSON object
after the value of `_output`.

Now, if the result only contains an `_option` key and the value for that
key is an empty string or whitespace, the human outputter will display a
message stating the action completed successfully with no result.