(GH-1493) Validate metadata when running or showing tasks
We now validate metadata during the construction of a Bolt::Task object,
which means a warning will be logged when running or showing a task if
that task contains metadata fields that are unknown to Bolt.

In some cases, an unknown metadata field may signal a typo or that
certain keys (such as parameter specifications) are in the wrong place
in the file. In others, it may mean that new features have been added to
the task spec that the current version of Bolt doesn't understand and
may not do the right thing with. In both cases, it's helpful to alert
the user to a potential problem.