(GH-1208) Bump puppetlabs-puppet_agent to 3.0.0
This bumps the version of the `puppet_agent` module to 3.0.0, which
includes changes to the `puppet_agent::install` task that are part of
this ticket.

This also updates a spec test to install the `latest` version of the
puppet-agent package, as the `puppet_agent::install` task no longer
assumes `latest` if an agent is already installed.