(maint) Refactor/rename Bolt::Util#read_config_file
All config (inventory.yaml, bolt.yaml, etc...) should have a hash type (or be empty). It is possible to have valid YAML that is not a hash type. This commit re-names the `Bolt::Util#read_config_file` method to `read_yaml_hash`. This change also removes some of the complexity in the method by no longer accepting a list of "default" paths to check. Previously if a list of defaults was passed in actually finding a file to read became optional. Now a second util method `read_optional_yaml_hash` has been added which will return the results of `read_yaml_hash` if the file exists and an empty hash otherwise. This allows the caller to choose the appropriate method based on if the file is expected to exist or not.