(GH-1567) Remove inventory, group, and target v1
This commit removes the inventory, group, and target v1 classes and
renames the v2 classes. The classes have been renamed as follows:

- `Bolt::Inventory::Inventory2` => `Bolt::Inventory::Inventory`
- `Bolt::Inventory::Group2` => `Bolt::Inventory::Group`
- `Bolt::Target2` => `Bolt::Target`

Additional changes include the following:

- The `Bolt::Inventory` class now acts as an inventory factory, with
several methods for generating v2 inventory objects.
- The `Bolt::CLI#migrate_project` method now parses an inventory file
for `nodes` keys instead of checking the `version` field.
- Added a `Bolt::Target#from_hash` method to create a target from a
config hash, add it to the inventory, and return a `Bolt::Target`
object.